+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 New roles:
 - Alpha Werewolf: Can turn other players into werewolves
-
+- Fake Seer: she only has a 50% chance of guessing the actual player's role
 
 Minor additions and changes
 - Added the option to disable the quota from the LethalConfig menu